+++ src/interfaces/Props.ts
@@ -133,4 +133,5 @@ export interface DropdownProps {
     onSelectOption: (selectedOption: DropdownOption) => void;
     selectedOption?: DropdownOption | null;
     dropdownIcon?: ReactNode;
+    dropdownInput?: ReactNode;
 }
