+++ src/pages/demos/react-dropdowns/index.html
@@ -7,9 +7,9 @@
             content="width=device-width, initial-scale=1"
         />
 
-        <title>React/Typescript Starter Dropdown</title>
+        <title>React/Typescript Starter Dropdowns</title>
         <script
-            src="../../demos/react-dropdown/index.html.js"
+            src="../../demos/react-dropdowns/index.html.js"
             defer
         ></script>
