--- src/styles/components/_dropdown.scss
@@ -1,132 +1,0 @@
-.dropdown {
-    position: relative;
-    width: 100%;
-    cursor: pointer;
-    border-radius: 4px;
-    border-bottom: 3px solid var(--legacy-2022-secondary-light-light-3-c-4-c-3-c-3, #8c9498);
-    background: var(--legacy-2022-tertiary-light-ligth-1-e-8-eef-1, #e8eef1);
-
-    &.dropdown-icon {
-        border-bottom: none;
-        position: relative;
-        background: #fff;
-
-        &:hover {
-            border-bottom: none;
-        }
-
-        &.open,
-        .completed {
-            border-bottom: none;
-        }
-
-        &.completed {
-            border-bottom: none;
-        }
-
-        .options {
-            position: absolute;
-            width: max-content;
-            right: 24px;
-            background-color: white;
-            box-shadow: 0 2px 10px rgba(0, 0, 0, 0.15);
-            border-radius: 4px;
-            padding: 10px 0;
-
-            .option {
-                padding: 10px 24px;
-                cursor: pointer;
-                &:hover {
-                    background-color: #f5f5f5;
-                }
-            }
-        }
-    }
-
-    &:hover {
-        border-bottom: 3px solid #0096c8;
-    }
-
-    &.open,
-    .completed {
-        border-bottom: 3px solid var(--legacy-2022-tertiary-base-005876, #005876);
-    }
-
-    &.completed {
-        border-bottom: 3px solid var(--legacy-2022-tertiary-base-005876, #005876);
-    }
-
-    .selected-option {
-        padding: 16px;
-        width: 100%;
-        color: var(--secondary-dark-1);
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-    }
-
-    .options {
-        position: absolute;
-        top: 100%;
-        left: 0;
-        width: 100%;
-        max-height: 264px;
-        background-color: #fff;
-        border-top: none;
-        overflow-y: auto;
-        list-style: none;
-        padding: 12px;
-        margin: 8px 0 0 0;
-        z-index: 1;
-        border-radius: 4px;
-        box-shadow: 0px 8px 10px 1px rgba(0, 0, 0, 0.14), 0px 3px 14px 2px rgba(0, 0, 0, 0.12),
-            0px 5px 5px -3px rgba(0, 0, 0, 0.2);
-
-        .option {
-            padding: 12px;
-            position: relative;
-            line-height: 24px;
-
-            &:hover {
-                background-color: #e8eef1;
-            }
-            &.selected {
-                background-color: #005876;
-                color: #fff;
-
-                &::before {
-                    content: '\f00c';
-                    font-family: 'Font Awesome 6 Pro';
-                    font-weight: var(--font-weight--regular);
-                    position: absolute;
-                    top: 50%;
-                    right: 24px;
-                    transform: translateY(-50%);
-                    width: 14px;
-                    height: 10px;
-                }
-            }
-        }
-    }
-
-    &.open {
-        .options {
-            display: block;
-        }
-    }
-}
-
-.col--lg-4 {
-    p {
-        margin-bottom: 16px;
-    }
-}
-
-.dropdown-requried {
-    p {
-        &::after {
-            content: ' *';
-            color: var(--error-element);
-        }
-    }
-}
