+++ src/styles/components/_dropdowns.scss
@@ -1,0 +1,202 @@
+.dropdown {
+    width: 100%;
+}
+
+.bg--tertiary-base {
+    .dropdown {
+        .selected-option {
+            background: #fff;
+            color: var(--legacy-2022-secondary-dark-dark-1949494, #6d7a7f);
+        }
+
+        .option {
+            color: var(--legacy-2022-secondary-dark-dark-1949494, #6d7a7f);
+        }
+    }
+
+    .dropdown-input,
+    .dropdown-icon {
+        .option {
+            color: var(--legacy-2022-secondary-dark-dark-1949494, #6d7a7f);
+        }
+    }
+}
+
+.dropdown,
+.dropdown-input {
+    position: relative;
+    display: inline-block;
+    cursor: pointer;
+
+    .selected-option {
+        padding: 16px;
+        width: 100%;
+        color: var(--secondary-dark-1);
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+        border-radius: 4px;
+        border-bottom: 3px solid var(--legacy-2022-secondary-light-light-3-c-4-c-3-c-3, #8c9498);
+        background: var(--legacy-2022-tertiary-light-ligth-1-e-8-eef-1, #e8eef1);
+
+        &:hover {
+            border-bottom: 3px solid #0096c8;
+        }
+
+        &.open,
+        .completed {
+            border-bottom: 3px solid var(--tertiary-base);
+        }
+
+        &.completed {
+            .completed-option {
+                color: var(--tertiary-base);
+            }
+            border-bottom: 3px solid var(--tertiary-base);
+        }
+    }
+
+    .options {
+        position: absolute;
+        top: 100%;
+        left: 0;
+        width: 100%;
+        max-height: 264px;
+        background-color: #fff;
+        border-top: none;
+        overflow-y: auto;
+        list-style: none;
+        padding: 12px 0;
+        margin: 8px 0 0 0;
+        z-index: 1;
+        border-radius: 4px;
+        box-shadow: 0px 8px 10px 1px rgba(0, 0, 0, 0.14), 0px 3px 14px 2px rgba(0, 0, 0, 0.12),
+            0px 5px 5px -3px rgba(0, 0, 0, 0.2);
+
+        .option {
+            padding: 12px 0 12px 24px;
+            position: relative;
+            line-height: 24px;
+
+            &:hover {
+                background-color: #e8eef1;
+            }
+            &.selected {
+                background-color: #005876;
+                color: #fff;
+                display: flex;
+                justify-content: space-between;
+                align-items: center;
+
+                i {
+                    font-size: 16px;
+                    padding-right: 19px;
+                }
+            }
+        }
+    }
+}
+
+.col--lg-4 {
+    p {
+        margin-bottom: 16px;
+    }
+}
+
+h2 {
+    &.fs--text-center {
+        margin-bottom: var(--sect--top-pd--sm);
+    }
+}
+
+.dropdown-requried {
+    p {
+        &::after {
+            content: ' *';
+            color: var(--error-element);
+        }
+    }
+}
+
+.dropdown-icon-button {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-end;
+}
+
+.dropdown-input-button {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+.col {
+    &.fs--text-center {
+        margin-top: var(--sect--top-pd--sm);
+    }
+}
+
+.row {
+    .col--md-odd-3 {
+        margin-top: var(--sect--top-pd--sm);
+    }
+}
+
+.dropdown-icon {
+    position: relative;
+    display: inline-block;
+
+    i {
+        font-size: 24px;
+        margin-right: 0;
+        cursor: pointer;
+    }
+
+    .options {
+        position: absolute;
+        width: max-content;
+        top: 100%;
+        right: 0;
+        display: flex;
+        flex-direction: column;
+        background-color: #fff;
+        max-height: 264px;
+        border-top: none;
+        overflow-y: auto;
+        padding: 12px 0;
+        margin: 8px 0 0 0;
+        border: 1px solid #ccc;
+        border-radius: 4px;
+        list-style: none;
+        z-index: 1;
+        box-shadow: 0px 8px 10px 1px rgba(0, 0, 0, 0.14), 0px 3px 14px 2px rgba(0, 0, 0, 0.12),
+            0px 5px 5px -3px rgba(0, 0, 0, 0.2);
+
+        .option {
+            padding: 12px 0 12px 24px;
+            position: relative;
+            line-height: 24px;
+            padding-right: 24px;
+
+            i {
+                margin-right: 0;
+                margin-left: 24px;
+            }
+
+            &:hover {
+                background-color: #e8eef1;
+            }
+            &.selected {
+                background-color: #005876;
+                color: #fff;
+                display: flex;
+                justify-content: space-between;
+                align-items: center;
+
+                i {
+                    font-size: 16px;
+                }
+            }
+        }
+    }
+}
