+++ src/styles/components/_index.scss
@@ -15,4 +15,4 @@
 @forward '_forms';
 @forward '_modals';
 @forward '_accordions';
-@forward '_dropdown';
+@forward '_dropdowns';
